Comments server

+++ src/xmlwriter.cpp
@@ -9,7 +9,7 @@ std::string XmlWriter::getXML(){
 
 void XmlWriter::buildXMLAck(){
     using namespace rapidxml;
-    
+
     if(function == "authentication"){
         buildHeader();
         buildAckContext();
@@ -17,7 +17,7 @@ void XmlWriter::buildXMLAck(){
     }
 }
 
-void XmlWriter::buildXMLActuate(std::vector<double> &status){
+void XmlWriter::buildXMLActuate(std::vector<float> &status){
     using namespace rapidxml;
 
     if(function == "actuateBool"){
@@ -58,7 +58,7 @@ void XmlWriter::finalizeAnswerToSR(){
     doc->append_node(root_node);
 }
 
-void XmlWriter::addDataToAnswer(std::string type, std::string clientName, std::vector<double> data){
+void XmlWriter::addDataToAnswer(std::string type, std::string clientName, std::vector<float> data){
     using namespace rapidxml;
     xml_node<> *wemos_node = doc->allocate_node(node_element, "wemos");
     xml_node<> *clientName_node = doc->allocate_node(node_element, "name", clientName.c_str());
@@ -68,19 +68,21 @@ void XmlWriter::addDataToAnswer(std::string type, std::string clientName, std::v
     for (std::size_t i = 0; i < data.size(); i++)
     {
         std::string dataNumber("data" + std::to_string(i + 1));
-        xml_node<> *data_node = doc->allocate_node(node_element, doc->allocate_string(dataNumber.c_str()), doc->allocate_string(std::to_string((int)std::round(data[i])).c_str()));
+        xml_node<> *data_node = doc->allocate_node(node_element, doc->allocate_string(dataNumber.c_str()),
+                                                   doc->allocate_string(std::to_string((int)std::round(data[i])).c_str()));
         wemos_node->append_node(data_node);
     }
     wemosjes_node->append_node(wemos_node);
 }
 
-void XmlWriter::buildActuateContext(std::vector<double> &status){
+void XmlWriter::buildActuateContext(std::vector<float> &status){
     using namespace rapidxml;
     function_node = doc->allocate_node(node_element, "function", "actuateBool");
     context_node = doc->allocate_node(node_element, "context", 0);
     for (size_t i = 0; i < status.size(); i++){
         std::string dataNumber("data" + std::to_string(i + 1));
-        xml_node<> *dataNode = doc->allocate_node(node_element, doc->allocate_string(dataNumber.c_str()), doc->allocate_string(std::to_string((int)std::round(status[i])).c_str()));
+        xml_node<> *dataNode = doc->allocate_node(node_element, doc->allocate_string(dataNumber.c_str()),
+                                                  doc->allocate_string(std::to_string((int)std::round(status[i])).c_str()));
         context_node->append_node(dataNode);
     }
     root_node->append_node(function_node);
@@ -92,14 +94,14 @@ void XmlWriter::buildHeader(){
     doc = new xml_document<>;
     root_node = doc->allocate_node(node_element, "message");
     char *node_name = doc->allocate_string("header");
-    header_node = doc->allocate_node(node_element, node_name); 
+    header_node = doc->allocate_node(node_element, node_name);
     sender_node = doc->allocate_node(node_element, "sender", senderName.c_str());
     receiver_node = doc->allocate_node(node_element, "receiver", receiverName.c_str());
     header_node->append_node(sender_node);
     header_node->append_node(receiver_node);
     root_node->append_node(header_node);
 
-    
+
 }
 
 void XmlWriter::buildStatusRequestContext(){
